+++ piptools/_compat/pip_compat.py
@@ -27,6 +27,7 @@ RequirementSet = do_import('req.req_set', 'RequirementSet')
 user_cache_dir = do_import('utils.appdirs', 'user_cache_dir')
 FAVORITE_HASH = do_import('utils.hashes', 'FAVORITE_HASH')
 is_file_url = do_import('download', 'is_file_url')
+path_to_url = do_import('download', 'path_to_url')
 url_to_path = do_import('download', 'url_to_path')
 PackageFinder = do_import('index', 'PackageFinder')
 FormatControl = do_import('index', 'FormatControl')
@@ -35,6 +36,8 @@ Command = do_import('cli.base_command', 'Command', old_path='basecommand')
 cmdoptions = do_import('cli.cmdoptions', old_path='cmdoptions')
 get_installed_distributions = do_import('utils.misc', 'get_installed_distributions', old_path='utils')
 PyPI = do_import('models.index', 'PyPI')
+stdlib_pkgs = do_import('utils.compat', 'stdlib_pkgs', old_path='compat')
+DEV_PKGS = do_import('commands.freeze', 'DEV_PKGS')
 
 # pip 18.1 has refactored InstallRequirement constructors use by pip-tools.
 if pkg_resources.parse_version(pip.__version__) < pkg_resources.parse_version('18.1'):